refactor: small updates - create on major map - general code clean - test case updates
