+++ tests/messages/json/test_map.py
@@ -5,9 +5,24 @@ from unittest.mock import Mock
 import pytest
 
 from deebot_client.event_bus import EventBus
-from deebot_client.events.map import MapSetType
+from deebot_client.events import MajorMapEvent, MapSetType
 from deebot_client.message import HandlingState
-from deebot_client.messages.json import OnMapSetV2
+from deebot_client.messages.json import OnMajorMap, OnMapSetV2
+from tests.helpers import get_message_json, get_success_body
+from tests.messages import assert_message
+
+
+async def test_OnPos() -> None:
+    expected_mid = "199390082"
+    value = "1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,1295764014,3378526980,2963288214,2739565817,729228561,2452519304,1295764014,1295764014"
+
+    json = get_message_json(get_success_body({"mid": expected_mid, "value": value}))
+
+    assert_message(
+        OnMajorMap,
+        json,
+        MajorMapEvent(expected_mid, value.split(","), requested=True),
+    )
 
 
 @pytest.mark.parametrize(
